Move get_by_index to ACER codes (only used by ACER)
